Refactor code to match code style
