--- src/main/java/bagPacker/exception/emptyInputException.java
@@ -1,4 +1,0 @@
-package bagPacker.exception;
-
-public class emptyInputException extends Exception {
-}
