+++ src/main/java/bagpacker/exception/EmptyInputException.java
@@ -1,0 +1,4 @@
+package bagpacker.exception;
+
+public class EmptyInputException extends Exception {
+}
