+++ src/main/java/bagpacker/iohandler/Storage.java
@@ -1,4 +1,4 @@
-package bagPacker.ioHandler;
+package bagpacker.iohandler;
 /**
  * Storage class contains methods to save and retrieve saved packing list in txt file
  */
