+++ src/main/java/bagpacker/packingfunc/Commands.java
@@ -1,4 +1,4 @@
-package bagPacker.listFunction;
+package bagpacker.packingfunc;
 /**
  * Commands class used to execute given commands from user
  */
