+++ src/main/java/bagpacker/packingfunc/Item.java
@@ -1,4 +1,4 @@
-package bagPacker.listFunction;
+package bagpacker.packingfunc;
 /**
  * Item class used to store information about item
  * Items to be packed can be saved in PackingList class
