+++ src/main/java/bagpacker/packingfunc/PackingList.java
@@ -1,4 +1,4 @@
-package bagPacker.listFunction;
+package bagpacker.packingfunc;
 /**
  * PackingList class contains methods to manipulate the users packing list
  */
