+++ src/test/java/bagpacker/iohandlertest/ParserTest.java
@@ -1,4 +1,4 @@
-package bagPacker.ioHandlerTest;
+package bagpacker.iohandlertest;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
