+++ src/test/java/bagpacker/packingfunctest/ItemTest.java
@@ -1,13 +1,13 @@
-package bagPacker.listFunctionTest;
+package bagpacker.packingfunctest;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * PackingListTest to test methods in PackingList class
+ * ItemTest to test methods in Item Class
  */
-public class PackingListTest {
+public class ItemTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
